Refactor withChildrenMock > withPropMock for simpler and more reusable usage
